fixed integer sizes

+++ contribs/timerwheel/timeout.c
@@ -331,8 +331,8 @@ TIMEOUT_PUBLIC void timeouts_del(struct timeouts *T, struct timeout *to) {
 
 		if (to->pending != &T->expired && TAILQ_EMPTY(to->pending)) {
 			ptrdiff_t index = to->pending - &T->wheel[0][0];
-			int wheel = index / WHEEL_LEN;
-			int slot = index % WHEEL_LEN;
+			int wheel = (int)index / WHEEL_LEN;
+			int slot = (int)index % WHEEL_LEN;
 
 			T->pending[wheel] &= ~(WHEEL_C(1) << slot);
 		}
